Refactor gif encoding to seperate file for cleaner code.
