+++ libs/takeScreenshot.js
@@ -1,9 +1,5 @@
 import puppeteer from 'puppeteer'
 
-
-// const url = process.argv[3]
-// const filename = process.argv[2]
-
 export async function takeSS(url, filename, device) {
   const browser = await puppeteer.launch()
   const page = await browser.newPage()
